(VMS-3693) - NX client crash
(VMS-3730) - Server should be able to migrate data having database from other cloud instance
(VMS-3744) - http pool UT, terminate test
Backed out changeset: 28337bdb9353
Internal: Merge fix.
Internal: Minor code improvements.
Internal: Minor code improvements; coding style fixes.
Internal: Refactoring: coding style fixed, unnecessary includes removed.
VMS-3258 RestAPI: Merge incomplete structures with DB contents
VMS-3632 Installer should not allow updating version with conflicting cloud instances
VMS-3638 Make API camera related functions consistent
VMS-3660 - User must be able to 'Disconnect System from Nx Cloud' from Nx client if there is no Internet on its servers.
VMS-3660 - User must be able to Disconnect from the cloud if no Internet on its servers
VMS-3664 tests and fixes
VMS-3730 - cloud host checking
VMS-3730 Server should be able to migrate data having database from other cloud instance
add missing file
compile error fix
fixed http client reconnect feature
linux compile fixes
network MultipleServerSocket addSocket fix and logs DEBUG1
vms-3052. Do not show Remove label next to the system storage.
vms-3052. Do not show Remove label next to the system storage. Merge heads.
vms-3052. User group validity check.
vms-3054. Limit runtime_actions record count.
vms-3054. Limit runtime_actions record count. Merge heads.
vms-3665. Valid storage is shown as 'unaccessible'in UI.
vms-3665. Valid storage is shown as 'unaccessible'in UI. Merge heads.
vms-4191. Resource status removal transaction.

<==== Mercurial data ====>
Branch: dev_3.0.0_gui

+++ common/src/plugins/resource/avi/thumbnails_stream_reader.cpp
@@ -124,7 +124,7 @@ void QnThumbnailsStreamReader::run()
 
         if (videoData)
             m_stat[videoData->channelNumber].onData(
-                videoData->dataSize(), videoData->flags & AV_PKT_FLAG_KEY);
+                (unsigned int) videoData->dataSize(), videoData->flags & AV_PKT_FLAG_KEY);
 
         putData(data);
     }
